requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.22.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-cloudinary-storage==0.3.0
django-crispy==1.8.1
django-heroku==0.3.1
django-smtp-ssl==1.0
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
pbr==5.5.0
Pillow==7.2.0
pipenv==2020.8.13
psycopg2==2.8.6
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
stevedore==3.2.1
urllib3==1.25.11
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
zipp==3.1.0
